#25 refactor: move directory ui -> layout
