+++ frontend/app/components/layout/Header.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import Link from 'next/link'
-import { Button } from './Button'
+import { Button } from '../ui/Button'
 
 export default function Header() {
   return (
