+++ frontend/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
-import Footer from './components/ui/Footer'
-import Header from './components/ui/Header'
+import Footer from './components/layout/Footer'
+import Header from './components/layout/Header'
 import './globals.css'
 
 export const metadata: Metadata = {
